fix: changed the appearance of the helper block
