+++ src/starting.ts
@@ -1,4 +1,5 @@
 import { renderHeroChoisePage } from "./render";
+import { wrapperBlokHelp, blockHelp, wrapperBlokHelpCopy, clickBlockHelp } from './type/helpGame';
 
 
 const container = document.querySelector('.container') as HTMLElement;
@@ -38,10 +39,12 @@ function welcome() {
   startBtn.addEventListener('click', onStartBtnClick);
 
   window.addEventListener('keydown', (event: KeyboardEvent) => {
-    if(event.key === 'Enter' && !isTypingStoryActive) {
+    if (event.key === 'Enter' && !isTypingStoryActive) {
       onStartBtnClick()
       isTypingStoryActive = true;
       window.removeEventListener('keydown', onSkipBtnClick)
+      const blockHelp = document.querySelector('.block-help') as HTMLElement
+      blockHelp.classList.add('active')
     }
     return;
   });
@@ -105,11 +108,22 @@ function displaySkipBtn() {
   skipBtn.innerHTML = 'Skip';
   container.append(skipBtn);
 
-  skipBtn.addEventListener('click', onSkipBtnClick)
+  skipBtn.addEventListener('click', () => {
+    onSkipBtnClick()
+    blockHelp()
+    const elementBlockHelp = document.querySelector('.block-help') as HTMLElement
+    elementBlockHelp.addEventListener('click', () => {
+    document.body.innerHTML = ''
+    wrapperBlokHelp();
+    wrapperBlokHelpCopy();
+    clickBlockHelp()
+})
+  })
 }
 
 function onSkipBtnClick() {
   container.remove();
   renderHeroChoisePage()
-
 }
+
+
